price update

+++ grocery/models.py
@@ -11,7 +11,7 @@ class Grocery(SoftDeletable):
     cover_image = models.ImageField()
     quantity = models.PositiveIntegerField()
     quantity_unit = models.ForeignKey(ConfigChoice, on_delete=models.CASCADE, related_name="quantity_unit")
-    price = models.DecimalField(decimal_places=2, max_digits=10)
+    price = models.IntegerField()
     category = models.ForeignKey(ConfigChoice, on_delete=models.CASCADE)
     added_by = models.ForeignKey(User, on_delete=models.CASCADE)
     created_at = models.DateTimeField(auto_now_add=True)
